+++ planners-search/src/numeric/evaluation/domain_abstractions/domain_abstraction_factory.rs
@@ -205,7 +205,7 @@ pub struct AbstractDistanceTable {
 
 #[derive(Debug, Clone)]
 pub struct WildcardPlanResult {
-    // Per-step set of concrete operator IDs
+    // Per-step set of concrete operator IDs.
     pub wildcard_plan: Vec<Vec<usize>>,
     // Path of abstract state hashes (`len = steps+1`).
     pub abstract_state_hashes: Vec<usize>,
@@ -1005,15 +1005,15 @@ impl DomainAbstractionFactory {
         let mut distances: Vec<f64> = vec![f64::INFINITY; num_states];
         let mut generating_op_ids: Vec<Option<usize>> = vec![None; num_states];
 
-        let mut core_vars: Vec<u32> = Vec::new();
+        let mut core_vars: Vec<usize> = Vec::new();
         for (v, &dom) in self.domain_sizes.iter().enumerate() {
             if dom > 1 {
-                core_vars.push(v as u32);
+                core_vars.push(v);
             }
         }
         for (n, &dom) in numeric_domain_sizes.iter().enumerate() {
             if dom > 1 {
-                core_vars.push((num_props + n) as u32);
+                core_vars.push(num_props + n);
             }
         }
         core_vars.sort_unstable();
